Use 'not_null' pointers to harden code
We are using frequently a frunction returning a pointer to a struct and
then access data elements from this scruct via pointer dereferencing.
Instead of always having to check for the pointer to be not the nullptr
in or code before dereferencing, use 'not_null' pointers from the
Guidelines Support Library (GSL). For more information see
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rf-nullptr

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>